revise conclusion
Originally it said that the types guaranteed the correct and full
reversal of an expression. It is hard to make that precise, because even
though the types help, it is still very possible to just drop some
information (for instance not remove a variable assignment from the
store when reversing a function application)

Using a language with a strong type system helps in general programming,
but it is hard to point at specific things that are useful in reversing
specifically that are not trivial.